Fixes test_libvirtr spawn_with_network_info test
This patch fixes the LibvirtConnTestCase.test_spawn_with_network_info
test case that was reported as bug 1053572.

The new implementation of the test makes no assumptions about the raised
exceptions, and exercises the LibvirtDriver.spawn method completely.
The test has the network_info faked and the libvirt connection mocked out.

In addition. it also fixes the fake_network.FakeVIFDriver class used in
this test to return a LibvirtConfigGuestInterface instance with fake
values instead of just a dict. Also fixes some docstring inaccuracies
in the nova.libvirt.vif module that may have lead to the FakeVIFDrivers
wrong implementation before.

Fixes bug 1053572

Change-Id: Ie5b69a2269ab099267231329eae21bd308fef835