Keep flavor information in system_metadata
This patch copies the instance_type information into an
instance's system_metadata on creation. This provides a way
for us to recall the attributes used to create the instance
even after the type may have been deleted.

This adds db migration 153, which generates flavor information in
system_metadata for existing instances in the database, as well as
brings a test to make sure that works.

Change-Id: Iad9d640b44df5f3831de58f9e800bcf42bc0610e