New instance_actions and events table, model, and api
These tables will be a place to record actions taken on an instance, and
related underlying events.  This will provide a log of what has taken
place which can aid in debugging or just satisfy curiosity.  It will
also hold error information in the event that something exceptional
occurs during the action.  This is setting up the ability to report
error information without needing to set the VM to an ERROR state.

Blueprint instance-actions

Change-Id: I1f793abf0e4df0e4977850127fc41b4750ba2f4f