Fixes KeyError: 'sr_uuid' when booting from volume on xenapi
This change fixes the capability for a XenServer/XCP hypervisor
to boot instances from an iscsi volume running on Cinder.

This patch takes into account that Cinder can be configured
with an iscsi back-end, instead of a xensm one, where keys like
sr_uuid are not being specified. It works through the magic of
determining the SR, and the vdi associated with the iscsi target
and lun.

This patch also does a bit of clean-up of code/comments that no
longer apply.

Addresses bug #1087308

Change-Id: Ie7d65eeb965b3468e4407981788725e2b43bff5e