Add option to make exception format errors fatal.
Adds a new fatal_exception_format_errors config option which
defaults to False. This option is use to control how the base
NovaException class handles errors which can occur when it
formats error messages.

The motivation for this change is to be able to enable exception
format checking in our tests by setting
fatal_exception_format_errors=True.

Change-Id: Ie96261ed96bcede4a2b5ec5600cb93c15141a800