Fix incorrect use of context object
As in nova/availability_zones.py, the method get_host_availability_zone()
expects "context" as a module rather than a RequestContext object.
But for our preference, the argument "context" is better to be a
RequestContext object. So fix this by passing an context object into the
method.

* Fix the usage of argument "context"
* Add unittest for the module "availability_zones"

Fix bug lp:#1100598

Change-Id: I5886ba26131261d97be65f18e9492c40401c5a7f