Merge LibvirtOpenVswitchDriver class into LibvirtGenericVIFDriver
The LibvirtGenericVIFDriver class can use the 'vif_type' mapping
field to determine whether an OVS network configuration is
required, thus a dedicated driver for OVS is no longer required.

The LibvirtOpenVswitchDriver class functionality is merged into
LibvirtGenericVIFDriver. For backwards compatibility with
the Folsom release, the existing LibvirtOpenVswitchDriver class
is made to inherit from LibvirtGenericVIFDriver and directly
call the bridge specific setup APIs. This eases migration to
the new VIF impl during the Grizzly deployment lifecycle, with
an expectation that the LibvirtOpenVswitchDriver stub will be
deleted after the Havana release.

Blueprint: libvirt-vif-driver
Change-Id: Iac79e0880b6fbcca0fd8cc26b59d23e579b947fa
Signed-off-by: Daniel P. Berrange <berrange@redhat.com>