Fix missing key error in libvirt.driver
The libvirt disk mapping refactor change introduced
a bug in libvirt.driver.get_guest_config where we were
attempting to perform instance_update using a non-existent
key in the disk_mapping.

Given that we already parsed out the root_device_name and stored
it in a variable, and then verify that it's set in an if
statement before this method we should be just using what
we've already parsed.

Fixes bug: 1118829

Change-Id: Ifd80e67740d6f257ee537b68e9afe028273484d2