Fix baremetal migration skipping.
The `have_mysql` function would only check for 'openstack_ci_test' and
not 'openstack_baremetal_test' yet was being used for the
baremetal-migration tests.

This resulted in tests to run that should have been skipped.

The solution is to have the test classes to define their configuraiton
requirements and then to pass these into the backend checking code
explicitly.

Change-Id: I436b5cfe4cace04e7bcccc2bed75b34948fe0272