Add MultipleCreate template and fix conflict with other templates
This patch creates a dedicated xml template to the os-multiple-create extension
when using return_reservation_id=True in the request. The previous code, with
all extensions loaded, conflicted with security_groups extension.

There is also a change in the response due to the new Template, but the only
information removed is an empty metadata field (which made the previous xml
output even more inconsisten with json). The consistency between json and xml
will be treated in the next api version.

DocImpact

Fixes bug 1149526

Change-Id: I3ca0129086ff1df3a8faa2b3b8bd959c9b74103c