xenapi: fix console for rescued instance
Check if the instance is in vm_state RESCUED and if so
change the name-label to reflect that. Name label is used
to get vm ref and currently the original instance is being
referenced even when in rescue mode. If rescue vm is not ready
yet, InstanceNotReady is raised.

Added tests cases for xenapi get_vnc_console for instance
and rescue mode.

Fixes bug #1157389

Change-Id: I5438103d517c40c1c87ddecd5f88b1487bfc0339