Do not test foreign keys with SQLite version < 3.7
Fixes bug 1158828

Even though SQLAlchemy is documented to support foreign keys (with
the pragma set) on SQLite >= 3.6.19, we're seeing consistent problems
with that support on RHEL 6.4 with SQLite 3.6.20.  It works fine on
Fedora and Ubuntu with SQLite 3.7.x.

To work around these problems, if we're using SQLite, test the SQLite
version inside the test that requires foreign key support,
and skip the test if it's < 3.7

Change-Id: I1c26f29371376b271a4e384ea7b5d2009651acc1