xenapi: remove auto_disk_config check during resize
This check was introduced by the following change:
7c5c8a743c21733120c85fdefb84b4357f5848d0

This is an alternative fix to ensure resize down
is not processed if the disk cannot be resized.

Note the failure now registers an instance action
and does not leave the VM in an error state.

In addition, this keeps the behavior that a failure
to resize up and instance during spawn will not
cause the operation to fail and put the VM into error.

fixes bug 1188135
fixes bug 1187934
fixes bug 1081225

Change-Id: I5b3dfdc7fcda26c55702dde6cdf191beee0c818c