Alphabetize v3 API extension entry point list
Alphabetize the v3 API extension point list in setup.cfg. With lots
of extensions being ported to V3 and an entry needed in setup.cfg for
each one and multiple people working on the ports if we always just
add an entry to the end of the list we are going to get a lot of merge
conflicts. To minimise the number of conflicts we'll try keeping the
list in alphabetical order and the number of conflicts should drop
as the list gets longer.

Change-Id: I386ec9bdb99660861d80f47923855e73f474a15f