Port images functionality to v3 API Part 2
The images functionality is moved for the V3 API from core to an
extension and is now accessed via /os-images instead of /images.

Consensus appears to be that we continue to need this
functionality for v3 because some deployers disable the public
image endpoint due security concerns and whether this will be
fixed in glance for Havana. Deployers can disable
this extension if they do not want it.

A further changeset will follow which changes the view builder
to correctly handle the lack of project id in the path with the
v3 API.

Partially implements blueprint nova-v3-api

Change-Id: I1237ad1ab5c58bb2d7cd9460b14666b4b97ef7fc