Port images metadata functionality to v3 API Part 2
The images metadata functionality is moved for the V3 API from
core to an extension and is now accessed via
/os-images/{image_id}/os-image-metadata instead of
/images/{image_id}/metadata.

Consensus appears to be that we continue to need this
functionality for v3 because some deployers disable the public
image endpoint due security concerns and whether this will be
fixed in glance for Havana. Deployers can disable this extension
if they do not want it.

Partially implements blueprint nova-v3-api

Change-Id: I13a6d049a29a452ad269dba7b59b03718ada2eab